+++ src/reanimated2/animation/animations.ts
@@ -177,413 +177,3 @@ export function withStartValue(startValue, animation) {
     return animation;
   });
 }
-
-// export function withStyleAnimation(styleAnimations) {
-//   'worklet';
-//   return defineAnimation({}, () => {
-//     'worklet';
-
-//     const onFrame = (animation, now) => {
-//       let stillGoing = false;
-//       Object.keys(styleAnimations).forEach((key) => {
-//         const currentAnimation = animation.styleAnimations[key];
-//         if (key === 'transform') {
-//           const transform = animation.styleAnimations.transform;
-//           for (let i = 0; i < transform.length; i++) {
-//             const type = Object.keys(transform[i])[0];
-//             const currentAnimation = transform[i][type];
-//             if (currentAnimation.finished) {
-//               continue;
-//             }
-//             const finished = currentAnimation.onFrame(currentAnimation, now);
-//             if (finished) {
-//               currentAnimation.finished = true;
-//               if (currentAnimation.callback) {
-//                 currentAnimation.callback(true);
-//               }
-//             } else {
-//               stillGoing = true;
-//             }
-//             animation.current.transform[i][type] = currentAnimation.current;
-//           }
-//         } else {
-//           if (!currentAnimation.finished) {
-//             const finished = currentAnimation.onFrame(currentAnimation, now);
-//             if (finished) {
-//               currentAnimation.finished = true;
-//               if (currentAnimation.callback) {
-//                 currentAnimation.callback(true);
-//               }
-//             } else {
-//               stillGoing = true;
-//             }
-//             animation.current[key] = currentAnimation.current;
-//           }
-//         }
-//       });
-//       return !stillGoing;
-//     };
-
-//     const onStart = (animation, value, now, previousAnimation) => {
-//       Object.keys(styleAnimations).forEach((key) => {
-//         if (key === 'transform') {
-//           animation.current.transform = [];
-//           const transform = styleAnimations.transform;
-//           const prevTransform = null;
-//           const valueTransform = value.transform;
-//           if (
-//             previousAnimation &&
-//             previousAnimation.styleAnimations &&
-//             previousAnimation.styleAnimations.transform
-//           ) {
-//             prevAnimation = previousAnimation.styleAnimations.transform;
-//           }
-
-//           for (let i = 0; i < transform.length; i++) {
-//             // duplication of code to avoid function calls
-//             let prevAnimation = null;
-//             const type = Object.keys(transform[i])[0];
-//             if (prevTransform && prevTransform.length > i) {
-//               const prevTransformStep = prevTransform[i];
-//               const prevType = Object.keys(prevTransformStep)[0];
-//               if (prevType === type) {
-//                 prevAnimation = prevTransformStep[prevType];
-//               }
-//             }
-
-//             let prevVal = 0;
-//             if (prevAnimation != null) {
-//               prevVal = prevAnimation.current;
-//             }
-//             if (
-//               valueTransform != null &&
-//               valueTransform.length > i &&
-//               valueTransform[i][type]
-//             ) {
-//               prevVal = valueTransform[i][type];
-//             }
-//             const obj = {};
-//             obj[type] = prevVal;
-//             animation.current.transform[i] = obj;
-//             let currentAnimation = transform[i][type];
-//             if (
-//               typeof currentAnimation !== 'object' &&
-//               !Array.isArray(currentAnimation)
-//             ) {
-//               currentAnimation = withTiming(currentAnimation, { duration: 0 });
-//               transform[i][type] = currentAnimation;
-//             }
-//             currentAnimation.onStart(
-//               currentAnimation,
-//               prevVal,
-//               now,
-//               prevAnimation
-//             );
-//           }
-//         } else {
-//           let prevAnimation = null;
-//           if (
-//             previousAnimation &&
-//             previousAnimation.styleAnimations &&
-//             previousAnimation.styleAnimations[key]
-//           ) {
-//             prevAnimation = previousAnimation.styleAnimations[key];
-//           }
-//           let prevVal = 0;
-//           if (prevAnimation != null) {
-//             prevVal = prevAnimation.current;
-//           }
-//           if (value[key]) {
-//             prevVal = value[key];
-//           }
-//           animation.current[key] = prevVal;
-//           let currentAnimation = animation.styleAnimations[key];
-//           if (
-//             typeof currentAnimation !== 'object' &&
-//             !Array.isArray(currentAnimation)
-//           ) {
-//             currentAnimation = withTiming(currentAnimation, { duration: 0 });
-//             animation.styleAnimations[key] = currentAnimation;
-//           }
-//           currentAnimation.onStart(
-//             currentAnimation,
-//             prevVal,
-//             now,
-//             prevAnimation
-//           );
-//         }
-//       });
-//     };
-
-//     const callback = (finished) => {
-//       if (!finished) {
-//         Object.keys(styleAnimations).forEach((key) => {
-//           const currentAnimation = styleAnimations[key];
-//           if (key === 'transform') {
-//             const transform = styleAnimations.transform;
-//             for (let i = 0; i < transform.length; i++) {
-//               const type = Object.keys(transform[i])[0];
-//               const currentAnimation = transform[i][type];
-//               if (currentAnimation.finished) {
-//                 continue;
-//               }
-//               if (currentAnimation.callback) {
-//                 currentAnimation.callback(false);
-//               }
-//             }
-//           } else {
-//             if (!currentAnimation.finished) {
-//               if (currentAnimation.callback) {
-//                 currentAnimation.callback(false);
-//               }
-//             }
-//           }
-//         });
-//       }
-//     };
-
-//     return {
-//       isHigherOrder: true,
-//       onFrame,
-//       onStart,
-//       current: {},
-//       styleAnimations,
-//       callback,
-//     };
-//   });
-// }
-
-export function withDelay(delayMs, _nextAnimation) {
-  'worklet';
-  return defineAnimation(_nextAnimation, () => {
-    'worklet';
-    const nextAnimation =
-      typeof _nextAnimation === 'function' ? _nextAnimation() : _nextAnimation;
-
-    function delay(animation, now) {
-      const { startTime, started, previousAnimation } = animation;
-
-      if (now - startTime > delayMs) {
-        if (!started) {
-          nextAnimation.onStart(
-            nextAnimation,
-            animation.current,
-            now,
-            previousAnimation
-          );
-          animation.previousAnimation = null;
-          animation.started = true;
-        }
-        const finished = nextAnimation.onFrame(nextAnimation, now);
-        animation.current = nextAnimation.current;
-        return finished;
-      } else if (previousAnimation) {
-        const finished = previousAnimation.onFrame(previousAnimation, now);
-        animation.current = previousAnimation.current;
-        if (finished) {
-          animation.previousAnimation = null;
-        }
-      }
-      return false;
-    }
-
-    function onStart(animation, value, now, previousAnimation) {
-      animation.startTime = now;
-      animation.started = false;
-      animation.current = value;
-      animation.previousAnimation = previousAnimation;
-    }
-
-    const callback = (finished) => {
-      if (nextAnimation.callback) {
-        nextAnimation.callback(finished);
-      }
-    };
-
-    return {
-      isHigherOrder: true,
-      onFrame: delay,
-      onStart,
-      current: nextAnimation.current,
-      callback,
-    };
-  });
-}
-
-export function withSequence(..._animations) {
-  'worklet';
-  return defineAnimation(_animations[0], () => {
-    'worklet';
-    const animations = _animations.map((a) => {
-      const result = typeof a === 'function' ? a() : a;
-      result.finished = false;
-      return result;
-    });
-    const firstAnimation = animations[0];
-
-    const callback = (finished) => {
-      if (finished) {
-        // we want to call the callback after every single animation
-        // not after all of them
-        return;
-      }
-      // this is going to be called only if sequence has been cancelled
-      animations.forEach((animation) => {
-        if (typeof animation.callback === 'function' && !animation.finished) {
-          animation.callback(finished);
-        }
-      });
-    };
-
-    function sequence(animation, now) {
-      const currentAnim = animations[animation.animationIndex];
-      const finished = currentAnim.onFrame(currentAnim, now);
-      animation.current = currentAnim.current;
-      if (finished) {
-        // we want to call the callback after every single animation
-        if (currentAnim.callback) {
-          currentAnim.callback(true /* finished */);
-        }
-        currentAnim.finished = true;
-        animation.animationIndex += 1;
-        if (animation.animationIndex < animations.length) {
-          const nextAnim = animations[animation.animationIndex];
-          nextAnim.onStart(nextAnim, currentAnim.current, now, currentAnim);
-          return false;
-        }
-        return true;
-      }
-      return false;
-    }
-
-    function onStart(animation, value, now, previousAnimation) {
-      if (animations.length === 1) {
-        throw Error(
-          'withSequence() animation require more than one animation as argument'
-        );
-      }
-      animation.animationIndex = 0;
-      if (previousAnimation === undefined) {
-        previousAnimation = animations[animations.length - 1];
-      }
-      firstAnimation.onStart(firstAnimation, value, now, previousAnimation);
-    }
-
-    return {
-      isHigherOrder: true,
-      onFrame: sequence,
-      onStart,
-      animationIndex: 0,
-      current: firstAnimation.current,
-      callback,
-    };
-  });
-}
-
-export function withRepeat(
-  _nextAnimation,
-  numberOfReps = 2,
-  reverse = false,
-  callback
-) {
-  'worklet';
-
-  return defineAnimation(_nextAnimation, () => {
-    'worklet';
-
-    const nextAnimation =
-      typeof _nextAnimation === 'function' ? _nextAnimation() : _nextAnimation;
-
-    function repeat(animation, now) {
-      const finished = nextAnimation.onFrame(nextAnimation, now);
-      animation.current = nextAnimation.current;
-      if (finished) {
-        animation.reps += 1;
-        // call inner animation's callback on every repetition
-        // as the second argument the animation's current value is passed
-        if (nextAnimation.callback) {
-          nextAnimation.callback(true /* finished */, animation.current);
-        }
-        if (numberOfReps > 0 && animation.reps >= numberOfReps) {
-          return true;
-        }
-
-        const startValue = reverse
-          ? nextAnimation.current
-          : animation.startValue;
-        if (reverse) {
-          nextAnimation.toValue = animation.startValue;
-          animation.startValue = startValue;
-        }
-        nextAnimation.onStart(
-          nextAnimation,
-          startValue,
-          now,
-          nextAnimation.previousAnimation
-        );
-        return false;
-      }
-      return false;
-    }
-
-    const repCallback = (finished) => {
-      if (callback) {
-        callback(finished);
-      }
-      // when cancelled call inner animation's callback
-      if (!finished && nextAnimation.callback) {
-        nextAnimation.callback(false /* finished */);
-      }
-    };
-
-    function onStart(animation, value, now, previousAnimation) {
-      animation.startValue = value;
-      animation.reps = 0;
-      nextAnimation.onStart(nextAnimation, value, now, previousAnimation);
-    }
-
-    return {
-      isHigherOrder: true,
-      onFrame: repeat,
-      onStart,
-      reps: 0,
-      current: nextAnimation.current,
-      callback: repCallback,
-    };
-  });
-}
-
-/* Deprecated section, kept for backward compatibility. Will be removed soon */
-export function delay(delayMs, _nextAnimation) {
-  'worklet';
-  console.warn('Method `delay` is deprecated. Please use `withDelay` instead');
-  return withDelay(delayMs, _nextAnimation);
-}
-
-export function repeat(
-  _nextAnimation,
-  numberOfReps = 2,
-  reverse = false,
-  callback
-) {
-  'worklet';
-  console.warn(
-    'Method `repeat` is deprecated. Please use `withRepeat` instead'
-  );
-  return withRepeat(_nextAnimation, numberOfReps, reverse, callback);
-}
-
-export function loop(nextAnimation, numberOfLoops = 1) {
-  'worklet';
-  console.warn('Method `loop` is deprecated. Please use `withRepeat` instead');
-  return repeat(nextAnimation, Math.round(numberOfLoops * 2), true);
-}
-
-export function sequence(..._animations) {
-  'worklet';
-  console.warn(
-    'Method `sequence` is deprecated. Please use `withSequence` instead'
-  );
-  return withSequence(..._animations);
-}
-/* Deprecated section end */
