+++ src/reanimated2/animation/index.ts
@@ -4,3 +4,6 @@ export * from './commonTypes';
 export * from './timing';
 export * from './spring';
 export * from './decay';
+export * from './delay';
+export * from './repeat';
+export * from './sequence';
